refractoring[DrivSPI]: delete redundancy code

+++ Src/U_Queue.c
@@ -12,9 +12,9 @@ int isEmpty(const Queue *const p) { return len(p) == 0; }
 int isFull(const Queue *const p) { return len(p) == QUEUE_CAPACITY; }
 
 // TODO: ERRO Handling
-uint8_t pop(Queue *const p) { return p->buffer[p->rcntr++ & QUEUE_MASK]; }
+uint32_t pop(Queue *const p) { return p->buffer[p->rcntr++ & QUEUE_MASK]; }
 // TODO: ERRO Handling
-void push(Queue *const p, uint8_t data) {
+void push(Queue *const p, uint32_t data) {
   p->buffer[p->wcntr++ & QUEUE_MASK] = data;
 }
